requirements.txt
# GUI
customtkinter>=5.2.0
Pillow>=10.0.0

# Utils
python-dotenv>=1.0.0
requests>=2.31.0

# EXE Build (opsiyonel)
# pyinstaller>=6.0.0